refactor: rearranged settings classes
